pyproject.toml
[tool.poetry]
name = "yandex-cloud-billing-analysis-bot"
version = "1.0.0"
description = ""
authors = ["Leonid Grishenkov <grishenkovleonid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "2.2.3"
boto3 = "1.35.34"
python-dotenv = "1.0.0"
python-telegram-bot = "21.6"
jinja2 = "3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
